requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.54
psycopg2-binary==2.9.6
alembic==1.11.1
requests==2.31.0
apscheduler==3.10.1
prometheus-client==0.16.0
kafka-python==2.1.0
redis==5.0.10
pydantic==1.10.11
python-dotenv==1.0.1
structlog==23.1.0
redis==5.0.10
pytest==7.4.0
pytest-mock==3.11.0
requests-mock==1.11.0
alembic==1.11.1
